Update Banking System.py
Very first OOP project. Honestly could use a lot of refactoring which I'll get to when I have time (the nested if-statements are disgusting, I know). 

Perhaps I'll also implement a main menu of sorts to navigate through the functions within the terminal, but I'm happy with how this turned out for now ^^